+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+v2.17.0
+========
+* Introduced converter for `char` primitive and `Character` class 
+
 v2.16.0
 =======
+* Introduced `EqualDay` specification which allows finding all records within particular date (day), ignoring time.
 * Added ability to set custom `Locale` during resolver registration:
   ```java
   @Override